requirement.txt
asgiref==3.5.2
Django==4.0.4
python-decouple==3.6
sqlparse==0.4.2
tzdata==2022.1
django-versatileimagefield==2.2
django-tinymce==3.4.0
psycopg2-binary